+++ src/main/java/com/example/demo/service/UsuarioService.java
@@ -6,14 +6,14 @@ import com.example.demo.entity.Usuario;
 
 public interface UsuarioService {
 
-    public Usuario SearchNyId(int id);
+    public Usuario SearchNyId(Long id);
 
     public Collection<Usuario> SearchAll();
 
-    public void deleletebyid(int id);
+    public void deleletebyid(Long id);
 
     public void update(Usuario usuario);
 
     public void add(Usuario usuario);
-    
-} 
+
+}
